Fix error when esdoc-lint-plugin is disabled
This plugins initializes the linting results (`this._results`) to `null`.

If this plugin is enabled, `this._results` is set to a new array when the `onPublish` event is handled. But when the plugin is disabled, `this._results` will never change and will remain set to `null`. This will cause an error when Esdoc tries to iterate over the linting results in: `for (const result of this._results) `

This change checks if the plugin is disabled and returns early, before showing the results.