Fail gracefully when splitting by comma doesn't work.
I know it's an edge case, but one of our exported objects has a key with a comma in it (don't ask), and it causes this to fail. I'm ok excluding it from the generated documentation (like I said, huge edge case) so this makes it so it at least finishes the html publishing.