Fix issue on named default exports.
To improve the way ESDoc indexes default exports, one should be able to
name the default exports.

For example, this is valid ES6:

  // in some api.js file...
  class Api {...}

  // this is the falty declaration
  export default api = new Api(host);

But the current version of ESDOC fails:

  $ esdoc -c esdoc.json
  [...]
  /.../node_modules/esdoc-type-inference-plugin/src/Plugin.js:80
      if (node.declarations[0].init.type === 'NewExpression') {

This is due to the _inferenceVariable() function not expecting to
receive an AssignmentExpression.

This patch simply adds a check for the presence of declaration(s) on the
parsed note that removes this issue.